packages/analysis/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "psx-analysis"
version = "1.0.0"
description = "Shared technical and financial analysis logic for PSX stocks"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "pandas>=1.5.0",
    "numpy>=1.23.0",
    "pdfplumber>=0.9.0",
    "langchain>=1.0.0",
    "langchain-core>=1.0.0",
    "langchain-community>=0.2.0",
    "langgraph>=1.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "pylint>=2.15.0",
    "ruff>=0.1.0",
]

[tool.setuptools]
packages = ["psx_analysis"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt"]

